requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5
psycopg==3.1.12
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
validate-docbr==1.10.0
Python==3.10.12
